requirements/dev.txt
-r base.txt

appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
black==25.1.0
click==8.1.8
comm==0.2.2
coverage==7.8.0
debugpy==1.8.14
decorator==5.2.1
django-debug-toolbar==5.2.0
django-extensions==3.2.3
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
executing==2.2.0
factory_boy==3.3.3
Faker==37.1.0
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isort==5.13.2
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumbers==9.0.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==5.0.0
pytest-django==4.10.0
pytest-factoryboy==2.7.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
wcwidth==0.2.13
